requirements.txt
django
djangorestframework
requests

